src/wearable-simulator/requirements.txt
paho-mqtt==1.6.1
numpy>=1.24.3
python-dotenv==1.0.0
pandas>=1.5.0
scipy>=1.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
